pyproject.toml
[project]
name = "pdf-rag-application"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.128.0",
    "inngest>=0.4.22",
    "llama-index-core>=0.14.12",
    "llama-index-readers-file>=0.5.6",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "qdrant-client>=1.16.1",
    "streamlit>=1.50.0",
    "uvicorn>=0.39.0",
]
